Final hardening: fix 8 issues from adversarial verification
Confirmed by a multi-dimension review + skeptic-refute pass (3 false positives
were correctly rejected):

- FilesScreen (high): FileRow declared onLong but never wired it, so long-press
  Rename/Delete was completely dead. Switched to combinedClickable(onLongClick).
- ServerDetailScreen (high): the PENDING_PAYMENT "Pay now" banner opened an
  external web checkout but was not gated by purchasingEnabled — now hidden on
  release builds like every other payment surface (Play §8).
- SupportListViewModel (high): the only paginated VM lacking an epoch/job guard;
  a load-more racing a refresh could append stale tickets and desync the page
  counter. Added the epoch + cancellable loadMoreJob guard the siblings use.
- EnvelopeConverterFactory (med): {success:true,data:null} threw instead of
  decoding to null; void/no-content success responses now succeed (iOS parity).
  Added EnvelopeConverterTest covering null/absent/present data.
- PushTokenRegistrar (med): register/unregister raced — a sign-out could be
  overtaken by a stale register, leaving a token bound to a signed-out account.
  All token mutations now run inside registerMutex with an in-lock session
  re-check.
- NotificationsScreen (med): permission status was a one-shot read that went
  stale after returning from system settings; now refreshed on ON_RESUME.
- AdminConfig.Coupon (low): money minor-unit fields were Int (~$21M cap); now
  Long like every other *Minor field.
- ServersListScreen (low): "1 server need attention" → "1 server needs
  attention".

+++ app/src/main/java/gg/refx/android/data/model/AdminConfig.kt
@@ -25,9 +25,10 @@ data class Coupon(
     val code: String,
     val description: String? = null,
     val kind: CouponKind = CouponKind.UNKNOWN,
-    val value: Int = 0,
+    // Money minor units for FIXED coupons — Long like every other *Minor field (no $21M Int cap).
+    val value: Long = 0,
     val currency: String = "USD",
-    val minSubtotalMinor: Int? = null,
+    val minSubtotalMinor: Long? = null,
     val maxRedemptions: Int? = null,
     val timesRedeemed: Int = 0,
     val maxPerUser: Int? = null,
@@ -40,7 +41,7 @@ data class Coupon(
     val valueLabel: String
         get() = when (kind) {
             CouponKind.PERCENT -> "$value%"
-            CouponKind.FIXED -> Money(value.toLong(), currency).formatted
+            CouponKind.FIXED -> Money(value, currency).formatted
             CouponKind.UNKNOWN -> value.toString()
         }
 }
